Save local changes before merging
