+++ src/components/Nav.module.css
@@ -12,7 +12,6 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
-
   max-width: 1920px;
   margin: 0 auto;
 }
