+++ src/components/ProductBestList.module.css
@@ -2,6 +2,13 @@ body {
   color: #1f2937;
 }
 
+.text {
+  margin-top: 30px;
+  font-size: 20px;
+  font-weight: 700;
+  color: #111827;
+}
+
 .BestItem {
   display: flex;
   justify-content: space-between;
@@ -46,6 +53,14 @@ body {
 }
 
 @media screen and (max-width: 1200px) {
+  .text {
+    font-size: 20px;
+    font-weight: 700;
+    align-items: center;
+    margin-top: 33px;
+    color: #111827;
+  }
+
   .productBestImg {
     width: 343px;
     height: 343px;
