+++ src/components/ProductListHeader.module.css
@@ -1,0 +1,97 @@
+body {
+  color: #1f2937;
+  font-weight: 200px;
+}
+
+.text {
+  margin-top: 30px;
+  font-size: 20px;
+  font-weight: 700;
+  color: #111827;
+}
+
+
+.productHeader {
+  display: flex;
+  margin-top: 40px;
+  justify-content: space-between;
+}
+
+.productControls {
+  display: flex;
+}
+
+.productControls > input {
+  position: relative;
+  width: 290px;
+  height: 42px;
+  background-color: #f3f4f6;
+  border: none;
+  border-radius: 12px;
+  padding-left: 44px;
+  color: #9ca3af;
+  margin-right: 15px;
+  margin-top: 20px;
+}
+
+.productControlsIcon {
+  position: absolute;
+  color: #9ca3af;
+  z-index: 1;
+  padding-left: 18px;
+  margin-top: 34px;
+}
+
+.productControls > button {
+  background-color: #3692ff;
+  border: none;
+  width: 133px;
+  height: 42px;
+  border-radius: 8px;
+  color: #f3f4f6;
+  font-size: 16px;
+  font-weight: 600;
+  margin-right: 12px;
+  justify-content: center;
+  align-items: center;
+  margin-top: 20px;
+}
+
+@media screen and (max-width: 1200px) {
+  .text {
+    font-size: 20px;
+    font-weight: 700;
+    align-items: center;
+    margin-top: 33px;
+    color: #111827;
+  }
+  .productControls > input {
+    width: 200px;
+  }
+}
+@media screen and (max-width: 773px) {
+
+  .productHeader {
+    display: grid;
+    position: relative;
+    height: 108px;
+    margin-top: 10px;
+    margin-bottom: 12px;
+  }
+
+  .productControls {
+    position: relative;
+    position: absolute;
+    bottom: -18px;
+  }
+
+  .productHeader .productControlsButton {
+    position: absolute;
+    bottom: 60px;
+    left: 72%;
+  }
+
+  .productControls > input {
+    width: 256px;
+  }
+}
