+++ src/index.js
@@ -4,3 +4,4 @@ import Main from './Main.js';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
 root.render(<Main />);
+
